# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register a patient or doctor account",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Authenticate with password and optional TOTP",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/mfa/setup": {
      "post": {
        "operationId": "AuthController_setupMfa",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/mfa/verify": {
      "post": {
        "operationId": "AuthController_verifyMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
...
    "/v1/doctors": {
      "get": {
        "operationId": "DoctorsController_search",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "specialization",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minFeeCents",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxFeeCents",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "availableFrom",
            "required": false,
            "in": "query",
            "description": "Only doctors with an available slot at/after this UTC time",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "doctors"
        ]
      }
    },
    "/v1/doctors/{id}/availability": {
      "get": {
        "operationId": "DoctorsController_availability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-10-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-11-01T00:00:00.000Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "doctors"
        ]
      }
    },
    "/v1/doctors/me/availability": {
      "post": {
        "operationId": "DoctorsController_createAvailability",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "doctors"
        ]
      }
    },
    "/v1/doctors/me/availability/{id}/block": {
      "patch": {
        "operationId": "DoctorsController_blockAvailability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "doctors"
        ]
      }
    },
    "/v1/admin/doctors/{id}/activation": {
      "patch": {
        "operationId": "AdminDoctorsController_setActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDoctorActivationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/bookings": {
      "post": {
        "operationId": "BookingsController_book",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookConsultationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "bookings"
        ]
      }
    },
    "/v1/consultations": {
      "get": {
        "operationId": "ConsultationsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SCHEDULED",
                "IN_PROGRESS",
                "COMPLETED",
                "CANCELLED",
                "NO_SHOW"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-11-01T00:00:00.000Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "consultations"
        ]
      }
    },
    "/v1/consultations/{id}": {
      "get": {
        "operationId": "ConsultationsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "consultations"
        ]
      }
    },
    "/v1/consultations/{id}/status": {
      "patch": {
        "operationId": "ConsultationsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsultationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "consultations"
        ]
      }
    },
    "/v1/consultations/{id}/prescriptions": {
      "post": {
        "operationId": "ConsultationsController_prescribe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrescriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "consultations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Amrutam Telemedicine API",
    "description": "Secure, idempotent REST API for telemedicine workflows",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "patient@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "example": "Str0ng!Passphrase"
          },
          "fullName": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "role": {
            "type": "string",
            "enum": [
              "PATIENT",
              "DOCTOR"
            ],
            "default": "PATIENT"
          },
          "phone": {
            "type": "string",
            "example": "+919876543210"
          },
          "specialization": {
            "type": "string",
            "example": "Ayurveda"
          },
          "licenseNumber": {
            "type": "string",
            "example": "MED-12345"
          },
          "consultationFeeCents": {
            "type": "number",
            "example": 150000,
            "description": "Fee in minor currency units (paise)"
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "patient@example.com"
          },
          "password": {
            "type": "string"
          },
          "totpCode": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "MfaVerifyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "code"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateAvailabilityDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "example": "2026-10-01T09:00:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2026-10-01T09:30:00.000Z"
          }
        },
        "required": [
          "startsAt",
          "endsAt"
        ]
      },
      "UpdateDoctorActivationDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "description": "Approve or deactivate a verified doctor profile"
          }
        },
        "required": [
          "active"
        ]
      },
      "BookConsultationDto": {
        "type": "object",
        "properties": {
          "slotId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "minLength": 3,
            "maxLength": 2000
          }
        },
        "required": [
          "slotId",
          "reason"
        ]
      },
      "UpdateConsultationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW"
            ]
          },
          "expectedVersion": {
            "type": "number",
            "description": "Current resource version for optimistic concurrency"
          },
          "clinicalNotes": {
            "type": "string",
            "maxLength": 10000
          }
        },
        "required": [
          "status",
          "expectedVersion"
        ]
      },
      "MedicationDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePrescriptionDto": {
        "type": "object",
        "properties": {
          "medications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicationDto"
            }
          },
          "instructions": {
            "type": "string",
            "maxLength": 4000
          }
        },
        "required": [
          "medications"
        ]
      },
...
    }
  }
}